% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{cohort.gp.density}
\alias{cohort.gp.density}
\title{Displays the distribution of genotyping probes values in a cohort.}
\usage{
cohort.gp.density(RnB.set)
}
\arguments{
\item{RnB.set}{A \code{RnBSet} basic object for storing methylation array
data and experimental quality information (HM450K and
Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}
}
\value{
A \code{gg} plot of the results.
}
\description{
Displays the distribution of genotyping probes values in a cohort.
}
\examples{
# Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
RnBeads::rnb.options(identifiers.column = "barcode")
# Draw the genotyping probes values offset plot
cohort.gp.density(RnB.set = rnb.set)
}
\author{
Yoann Pageaud.
}
