% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{comp_RnBqc2PCA}
\alias{comp_RnBqc2PCA}
\title{Computes a PCA from an RnBSet on quality control probes intensities}
\usage{
comp_RnBqc2PCA(RnBSet)
}
\arguments{
\item{RnBSet}{An \code{RnBSet} basic object for storing methylation
array DNA methylation and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}
}
\value{
A \code{list} containing a prcomp object with all results from the
        PCA, and a data.table with all the RnBSet data.
}
\description{
Computes a PCA from an RnBSet on quality control probes intensities
}
\examples{
# Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
# Compute PCA on quality control probes intensities from the RnBSet
comp_RnBqc2PCA(RnBSet = rnb.set)     
}
\author{
Yoann Pageaud.
}
