% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{compute.intensity.ratio}
\alias{compute.intensity.ratio}
\title{Computes ratio of Cy3 & Cy5 intensities with associated color shades.}
\usage{
compute.intensity.ratio(DT.probe.ratio)
}
\arguments{
\item{DT.probe.ratio}{A \code{data.table} with 3 columns:
\itemize{
 \item{column 1 "Samples" must contain all sample IDs.}
 \item{column 2 "Cy3 intensity" must contain Cy3 intensity values.}
 \item{column 3 "Cy5 intensity" must contain Cy5 intensity values.}
}}
}
\value{
A \code{data.table} with 7 columns:
\itemize{
 \item{column 1 "Samples" contains all sample IDs.}
 \item{column 2 "Cy3 intensity" contains Cy3 intensity values.}
 \item{column 3 "Cy5 intensity" contains Cy5 intensity values.}
 \item{column 4 "Intensity ratio" contains Cy3/Cy5 or Cy5/Cy3 ratio values.}
 \item{column 5 "Ratio.type" specify if the ratio is Cy3/Cy5 or Cy5/Cy3.}
 \item{column 6 "color.ratio" contains colors associated to ratio values.}
 \item{column 7 "round.ratio" contains the rounded ratio values.}
}
}
\description{
Computes ratio of Cy3 & Cy5 intensities with associated color shades.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
