% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{devscore.heatmaps}
\alias{devscore.heatmaps}
\title{Plots QC deviation heatmaps based on samples fluorescence deviation score.}
\usage{
devscore.heatmaps(
  RnBSet,
  target,
  samples = NULL,
  axis.text.y.left = ggplot2::element_text(size = 12, colour = "black"),
  axis.text.x = ggplot2::element_text(size = 12, angle = -45, hjust = 0, vjust = 0.5,
    colour = "black"),
  ncores = 1
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation
array data and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation
 array data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{target}{A \code{character} string specifying the QC target
type. Each 'target' matches a specific step in
Illumina array methods. Supported values:
target = c("Bisulfite Conversion I",
"Bisulfite Conversion II", "Extension",
"Hybridization", "Negative", "Non-polymorphic",
"Norm A", "Norm C", "Norm G", "Norm T",
"Specificity I", "Specificity II", "Staining",
"Target Removal").}

\item{samples}{A \code{character} vector specifying the samples to
include for the deviation score calculation. You can
catch the sample IDs you wish to evaluate running
\code{RnBSet@pheno[,1]}}

\item{axis.text.y.left}{An \code{element_text} object to setup right Y axis
text (Default: axis.text.y.right =
element_text(size = 12, color = "black")).}

\item{axis.text.x}{An \code{element_text} object to setup X axis text
(Default: axis.text.x = element_text(size = 12,
angle = -45, hjust = 0, vjust = 0.5,
color = "black")).}

\item{ncores}{An \code{integer} specifying the number of cores or
threads to be used for parallel processing.}
}
\value{
A \code{list} of 2 heatmaps as grid objects (grobs):
        \itemize{
         \item{one heatmap for probes deviation scores in the green channel.}
         \item{one heatmap for probes deviation scores in the red channel.}
        }
}
\description{
Plots QC deviation heatmaps based on samples fluorescence deviation score.
}
\details{
For more information about how the fluorescence deviation score is
         calculated, please refer to the details section of
         \link{devscore.fluo}.
}
\examples{
#Create an RnBSet for HM450K data
library(RnBeads)
idat.dir <- "~/data/my_idat_dir/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(
  data.source = data.source, data.type = "idat.dir")
#Draw deviation score heatmaps for green and red channels
dev.heatmaps <- devscore.heatmaps(
  RnBSet = rnb.set, samples = c("13169", "13947", "14312", "14359"),
  target = "Hybridization")
#One can then access both heatmaps from the list dev.heatmaps
}
\author{
Yoann Pageaud.
}
