% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_fix.R
\name{draw_key_boxplot2}
\alias{draw_key_boxplot2}
\title{Fixes display issues of geom_boxplot and geom_crossbar legend keys.}
\usage{
draw_key_boxplot2(data, params, size)
}
\description{
Fixes display issues of geom_boxplot and geom_crossbar legend keys.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
