% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{get.expected.intensity}
\alias{get.expected.intensity}
\title{Provides expected intensity for a given HM450K probe ID.}
\usage{
get.expected.intensity(DT.QC.meta, probe.id, channel.names)
}
\arguments{
\item{DT.QC.meta}{A \code{data.table} with HM450K quality control metadata
obtained using \link{load.HM450K.QC.meta}.}

\item{probe.id}{A \code{character} string matching a HM450K probe ID
listed in 'DT.QC.meta'.}

\item{channel.names}{A \code{character} vector specifying the names of the 2
color channels to use for creating the data.table of
expected intensity.}
}
\value{
A \code{data.table} with a 'Channel' column and an
        'Expected intensity' column providing the probe intensity expected
        for green and red channels.
}
\description{
Provides expected intensity for a given HM450K probe ID.
}
\references{

}
\author{
Yoann Pageaud.
}
