% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{get_IDATs_runinfo}
\alias{get_IDATs_runinfo}
\title{Retrieves runinfo data from a methylation array sample's IDAT files.}
\usage{
get_IDATs_runinfo(sentrix_barcode, IDATs_dir, data_format = "both")
}
\arguments{
\item{sentrix_barcode}{A \code{character} string to specify the sample from
which you wish to retrieve the runinfo (from both Red
& Green IDAT files).}

\item{IDATs_dir}{A \code{character} specifying the directory where IDAT
files are stored that will be search recursively for
the files matching the sentrix barcode.}

\item{data_format}{A \code{character} string to specify how much
information you want from runinfo:
\itemize{
 \item{'full': to get all details available from
 the sample's IDAT files runinfo.}
 \item{'short': to get a one line summary of runinfo
 data from the sample's IDAT files (Scan year,
 Decoding tool version & Scanning tool version).}
 \item{'both': the default value, to get both the full
 details and the one line summary of runinfo data from
 the sample's IDAT files in 2 separate data.tables.}
}}
}
\value{
A \code{data.table} or a \code{list} of data.tables with the sample's
        runinfo data.
}
\description{
Retrieves runinfo data from a methylation array sample's IDAT files.
}
\examples{
# Set IDAT directory
idat.dir <- system.file("extdata", "idat", package = "IlluminaDataTestFiles")
# Get runinfo data from a IDAT file name from within the directory
get_IDATs_runinfo(sentrix_barcode = "4019585376_B_Red", IDATs_dir = idat.dir)
}
\references{
ML Smith, KA Baggerly, H Bengtsson, ME Ritchie & KD Hansen.
            illuminaio: An open source IDAT parsing tool for Illumina
            microarrays, F1000Research, (2013) 2:264.
}
\author{
Yoann Pageaud.
}
