% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{get_platform}
\alias{get_platform}
\title{Detects platform used to generate data in the RnBSet.}
\usage{
get_platform(RnBSet)
}
\arguments{
\item{RnBSet}{An \code{RnBSet} basic object for storing methylation array DNA
methylation and experimental quality information (Bisulfite
data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array data
 in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}
}
\value{
A \code{character} string:
        \itemize{
         \item{"HM450K" if the RnBSet contains Human Methylation 450K data}
         \item{"MethylationEPIC" if the RnBSet contains MethylationEPIC data}
        }
}
\description{
Detects platform used to generate data in the RnBSet.
}
\examples{
#Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
#Check platform used to generate the dataset.
get_platform(rnb.set)
}
\author{
Yoann Pageaud.
}
