% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{gp_density_map}
\alias{gp_density_map}
\title{Displays the distribution of genotyping probes values in a cohort.}
\usage{
gp_density_map(RnB.set, show.sampleIDs = TRUE, draw = TRUE)
}
\arguments{
\item{RnB.set}{A \code{RnBSet} basic object for storing methylation array
data and experimental quality information (HM450K and
Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{show.sampleIDs}{A \code{logical} to specify whether the sample IDs
should be displayed on the Y axis
(Default: show.sampleIDs = TRUE) or hidden
(show.sampleIDs = FALSE).}

\item{draw}{A \code{logical} specifying whether the result plot
should be automatically displayed
(Default: draw = TRUE) or not (draw = FALSE).}
}
\value{
A \code{gg} plot of the results.
}
\description{
Displays the distribution of genotyping probes values in a cohort.
}
\examples{
# Create an RnBSet for MethylationEPIC data (May take up to 1 minute)
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
RnBeads::rnb.options(identifiers.column = "barcode")
# Draw the genotyping probes values offset plot
gp_density <- gp_density_map(RnB.set = rnb.set)
}
\author{
Yoann Pageaud.
}
