% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{load.EPIC.QC.meta}
\alias{load.EPIC.QC.meta}
\title{Loads MethylationEPIC QC metadata as a data.table.}
\usage{
load.EPIC.QC.meta()
}
\value{
A \code{data.table} with MethylationEPIC quality control metadata.
}
\description{
Loads MethylationEPIC QC metadata as a data.table.
}
\examples{
DT.QC.meta <- load.EPIC.QC.meta()
}
\references{
Assenov Y. et al., Comprehensive analysis of DNA methylation data
            with RnBeads.
}
\author{
Yoann Pageaud.
}
