% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{load.HM450K.QC.meta}
\alias{load.HM450K.QC.meta}
\title{Loads HM450K QC metadata as a data.table.}
\usage{
load.HM450K.QC.meta()
}
\value{
A \code{data.table} with HM450K quality control metadata.
}
\description{
Loads HM450K QC metadata as a data.table.
}
\examples{
DT.QC.meta <- load.HM450K.QC.meta()
}
\references{
Assenov Y. et al., Comprehensive analysis of DNA methylation data
            with RnBeads.
}
\author{
Yoann Pageaud.
}
