% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{load.HM450K.QC.theme}
\alias{load.HM450K.QC.theme}
\title{Loads QC ggplot2 default theme.}
\usage{
load.HM450K.QC.theme()
}
\value{
A ggplot2 \code{theme} object to be used as a default theme for QC
        plots.
}
\description{
Loads QC ggplot2 default theme.
}
\references{

}
\author{
Yoann Pageaud.
}
\keyword{internal}
