% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{load.metharray.QC.theme}
\alias{load.metharray.QC.theme}
\title{Loads QC ggplot2 default theme.}
\usage{
load.metharray.QC.theme()
}
\value{
A ggplot2 \code{theme} object used as a default theme for QC
        plots.
}
\description{
Loads QC ggplot2 default theme.
}
\examples{
theme_qc <- methview.qc::load.metharray.QC.theme()
}
\references{

}
\author{
Yoann Pageaud.
}
\keyword{internal}
