% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{load_metharray_QC_meta}
\alias{load_metharray_QC_meta}
\title{Loads methylation array QC metadata as a data.table.}
\usage{
load_metharray_QC_meta(array.meta)
}
\arguments{
\item{array.meta}{A \code{character} string specifying the array type to load
quality control metadata from
(Supported: array.meta = c("controls450", "controlsEPIC")).}
}
\value{
A \code{data.table} with HM450K or MethylationEPIC quality control
        metadata depending on the QC metadata requested.
}
\description{
Loads methylation array QC metadata as a data.table.
}
\examples{
#Load quality control metadata for Human Methylation 450K array
DT.QC.meta <- load_metharray_QC_meta(array.meta = "controls450")
}
\references{
Assenov Y. et al., Comprehensive analysis of DNA methylation data
            with RnBeads.
}
\author{
Yoann Pageaud.
}
