% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{load_metharray_QCtheme}
\alias{load_metharray_QCtheme}
\title{Loads QC ggplot2 default theme.}
\usage{
load_metharray_QCtheme()
}
\value{
A ggplot2 \code{theme} object used as a default theme for QC
        plots.
}
\description{
Loads QC ggplot2 default theme.
}
\examples{
theme_qc <- methview.qc:::load_metharray_QCtheme()
}
\author{
Yoann Pageaud.
}
\keyword{internal}
