% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{merge.QC.intensities.and.meta}
\alias{merge.QC.intensities.and.meta}
\title{Merges red and green channels intensities with QC probes metadata.}
\usage{
\method{merge}{QC.intensities.and.meta}(RnBSet, DT.QC.meta)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing HM450K DNA
methylation and experimental quality information
(MethylationEPIC & Bisulfite data not supported).}

\item{DT.QC.meta}{A \code{data.table} with HM450K quality control metadata
obtained using \link{load.HM450K.QC.meta}.}
}
\value{
A \code{data.table} list matching QC metadata with green channel and
        red channel intensities.
}
\description{
Merges red and green channels intensities with QC probes metadata.
}
\references{

}
\author{
Yoann Pageaud.
}
