% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_fix.R
\name{null_default}
\alias{null_default}
\title{Checks if x is null and if so uses y by default.}
\usage{
null_default(x, y)
}
\arguments{
\item{x}{A value to be checked for being NULL or not.}

\item{y}{Another value to be returned if x is NULL.}
}
\value{
A value that is not NULL.
}
\description{
Checks if x is null and if so uses y by default.
}
\references{
Henry L, Wickham H (2022). rlang: Functions for Base Types and
            Core R and 'Tidyverse' Features. https://rlang.r-lib.org,
            https://github.com/r-lib/rlang.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
