% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot.array.QC.probe}
\alias{plot.array.QC.probe}
\title{Plots fluorescence intensities barplots for a single QC methylation array probe.}
\usage{
\method{plot}{array.QC.probe}(array.type = "HM450K", probe.ID, QC.data, DT.QC.meta, cohort = "RnBSet")
}
\arguments{
\item{array.type}{A \code{character} string specifying the type of array used
to generate the data (Default: array.type = "HM450K";
Supported: array.type = c("HM450K", "EPIC")).}

\item{probe.ID}{A \code{character} string specifying the ID of a
methylation array quality control probe.}

\item{QC.data}{A \code{data.table} list matching QC metadata with green
channel and red channel intensities, obtained with the
function \link{merge.QC.intensities.and.meta}.}

\item{DT.QC.meta}{A \code{data.table} with methylation array quality control
metadata obtained with the function
\link{load.metharray.QC.meta}.}

\item{cohort}{A \code{character} string to specify the name of the cohort
to be displayed as part of the plot title
(Default: cohort = "RnBSet").}
}
\value{
A \code{gtable} barplot of the QC probe fluorescence
        intensities.
}
\description{
Plots fluorescence intensities barplots for a single QC methylation array probe.
}
\examples{
#Create an RnBSet for MethylationEPIC data
library(RnBeads)
idat.dir <- "~/data/MethylationEPIC/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(data.source = data.source, data.type = "idat.dir")
rnb.options(identifiers.column = "barcode")
#Create the data.table with quality control metadata
dt.meta <- load.metharray.QC.meta(array.meta = "controlsEPIC")
# Merge red and green channels intensities with QC metadata
dt.mrg <- merge.QC.intensities.and.meta(RnBSet = rnb.set, DT.QC.meta = dt.meta)
#Draw probe specific QC plot for QC probe "21630339"
probe.plot <- HM450.QCView:::plot.array.QC.probe(
  array.type = "EPIC", probe.ID = "21630339", QC.data = dt.mrg,
  DT.QC.meta = dt.meta)
#Save plot in a PDF file
ggsave(filename = "QCprobe_21630339.pdf", plot = probe.plot, device = "pdf",
       path = "~/")
}
\author{
Yoann Pageaud.
}
