% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot.HM450K.QC.probe}
\alias{plot.HM450K.QC.probe}
\title{Plots fluorescence intensities barplots for a single QC HM450K probe.}
\usage{
\method{plot}{HM450K.QC.probe}(probe.ID, QC.data, DT.QC.meta, cohort = "RnBSet")
}
\arguments{
\item{probe.ID}{A \code{character} string specifying the ID of an HM450K
quality control probe.}

\item{QC.data}{A \code{data.table} list matching QC metadata with green
channel and red channel intensities, obtained with the
function \link{merge.QC.intensities.and.meta}.}

\item{DT.QC.meta}{A \code{data.table} with HM450K quality control metadata
obtained with the function \link{load.HM450K.QC.meta}.}

\item{cohort}{A \code{character} string to specify the name of the cohort
to be displayed as part of the plot title
(Default: cohort = "RnBSet").}
}
\value{
A \code{gtable} barplot of the QC HM450K probe fluorescence
        intensities.
}
\description{
Plots fluorescence intensities barplots for a single QC HM450K probe.
}
\references{

}
\author{
Yoann Pageaud.
}
