% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot.HM450K.QC.target}
\alias{plot.HM450K.QC.target}
\title{Plots samples fluorescence intensities distribution for QC HM450K probes of a
specific target type.}
\usage{
\method{plot}{HM450K.QC.target}(
  array.type = "HM450K",
  target,
  QC.data,
  DT.QC.meta,
  cohort = "RnBSet",
  ncores = 1
)
}
\arguments{
\item{array.type}{A \code{character} string specifying the type of array used
to generate the data (Default: array.type = "HM450K";
Supported: array.type = c("HM450K", "EPIC")).}

\item{target}{A \code{character} string specifying the HM450K QC target
type. Each 'target' matches a specific step in Illumina array
methods. Supported values:
target = c("Bisulfite Conversion I",
"Bisulfite Conversion II", "Extension", "Hybridization",
"Negative", "Non-polymorphic", "Norm A", "Norm C", "Norm G",
"Norm T", "Specificity I", "Specificity II", "Staining",
"Target Removal").}

\item{QC.data}{A \code{data.table} list matching QC metadata with green
channel and red channel intensities, obtained with the
function \link{merge.QC.intensities.and.meta}.}

\item{DT.QC.meta}{A \code{data.table} with HM450K quality control metadata
obtained with the function \link{load.HM450K.QC.meta}.}

\item{cohort}{A \code{character} string to specify the name of the cohort
to be displayed as part of the plot title
(Default: cohort = "RnBSet").}

\item{ncores}{An \code{integer} to specify the number of cores/threads to
be used to parallel-compute probes intensities.}
}
\value{
A \code{gg} plot. The resulting plot depends of 2 parameters:
        \itemize{
         \item{The target type:
          \itemize{
           \item{If target = "Negative", the result is is a list of length 4,
                 containing 4 plots depicting fluorescence intensities of
                 negative probes in the red and green channels.}
           \item{If the target is of any other type, only 1 plot will be
                 returned. The resulting plot may still vary from one target
                 type to another.}
          }
         }
         \item{The number of samples considered:
          \itemize{
           \item{If 'QC.data' contains less than 5 samples, probes
                 fluorescence intensities will be displayed as dots, 1 dot
                 per sample.}
           \item{If 'QC.data' contains between 5 and 29 samples, probes
                 fluorescence intensities will be displayed as boxplots.}
           \item{If 'QC.data' contains 30 samples or more, probes
                 fluorescence intensities will be displayed as violin plots.}
          }
         }
        }
}
\description{
Plots samples fluorescence intensities distribution for QC HM450K probes of a
specific target type.
}
\references{

}
\author{
Yoann Pageaud.
}
