% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot_all_qc}
\alias{plot_all_qc}
\title{Draws and saves all quality control plots available in methview.qc}
\usage{
plot_all_qc(
  RnBSet,
  cohort = "RnBSet",
  save.dir,
  ncores = 1,
  include.gp = TRUE,
  include.ds = TRUE,
  include.ffpe = FALSE
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation array
data and experimental quality information (Bisulfite data
not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{cohort}{A \code{character} string to specify the name of the cohort
to be displayed as part of the plot title
(Default: cohort = "RnBSet").}

\item{save.dir}{A \code{character} string to specify the path where the
quality control plots should be saved.}

\item{ncores}{An \code{integer} to specify the number of cores/threads to
be used to parallel-compute probes intensities.}

\item{include.gp}{A \code{logical} to specify whether the genotyping probes
heatmap should be plotted too (include.gp = TRUE) or not
(include.gp = FALSE).\cr If you wish to customize your
genotyping probes heatmap, include.gp must be set to FALSE.
For more information see the details section.}

\item{include.ds}{A \code{logical} to specify whether the fluorescence
deviation heatmap should be plotted too (include.ds = TRUE)
or not (include.ds = FALSE).\cr If you wish to customize
your fluorescence deviatiom heatmap, include.ds must be set
to FALSE. For more information see the details section.}

\item{include.ffpe}{A \code{logical} to specify whether the FFPE negative
control probe should be plotted a second time separately
from the rest of the Negative QC probes (Warning: this is
an experimental function; for more information about the
FFPE negative control probe see
\link{plot_negative_FFPE}).}
}
\description{
Draws and saves all quality control plots available in methview.qc
}
\details{
The genotyping probes heatmap and the fluorescence deviation heatmap produced
by \link{plot_all_qc} when \code{include.gp = TRUE} and 
\code{include.ds = TRUE} respectively are automatic, non-custom heatmaps
created using \link{snp_heatmap} and \link{devscore.heatmap} respectively.
You can customize any of these heatmaps using directly \link{snp_heatmap} or
\link{devscore.heatmap} outside \link{plot_all_qc}. This way, you can provide
custom annotations to the top annotation bar, and personalized a lot more the
different components of your plots.
}
\examples{
#Create an RnBSet for MethylationEPIC data
library(RnBeads)
idat.dir <- "~/data/MethylationEPIC/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(data.source = data.source, data.type = "idat.dir")
rnb.options(identifiers.column = "barcode")
#Draw all plots from the quality control data of rnb.set
plot_all_qc(RnBSet = rnb.set, save.dir = "~/", ncores = 2)
}
\author{
Yoann Pageaud.
}
