% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot_array_QCprobe}
\alias{plot_array_QCprobe}
\title{Plots fluorescence intensities barplots for a single QC methylation array probe.}
\usage{
plot_array_QCprobe(
  array.type = "HM450K",
  probe.ID,
  QC.data,
  DT.QC.meta,
  cohort = "RnBSet",
  draw = TRUE
)
}
\arguments{
\item{array.type}{A \code{character} string specifying the type of array used
to generate the data (Default: array.type = "HM450K";
Supported: array.type = c("HM450K", "EPIC")).}

\item{probe.ID}{A \code{character} string specifying the ID of a
methylation array quality control probe. You can find probe
IDs in the data.table generated by
\link{load_metharray_QC_meta} in the first column "ID".}

\item{QC.data}{A \code{data.table} list matching QC metadata with green
channel and red channel intensities, obtained with the
function \link{mergeQC_intensities_and_meta}.}

\item{DT.QC.meta}{A \code{data.table} with methylation array quality control
metadata obtained with the function
\link{load_metharray_QC_meta}.}

\item{cohort}{A \code{character} string to specify the name of the cohort
to be displayed as part of the plot title
(Default: cohort = "RnBSet").}

\item{draw}{A \code{logical} specifying whether the result plot should
be automatically displayed (Default: draw = TRUE) or not
(draw = FALSE).}
}
\value{
A \code{gtable} barplot of the QC probe fluorescence
        intensities.
}
\description{
Plots fluorescence intensities barplots for a single QC methylation array probe.
}
\examples{
# Create an RnBSet for MethylationEPIC data (May take up to 1 minute)
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
RnBeads::rnb.options(identifiers.column = "barcode")
#Create the data.table with quality control metadata
dt.meta <- load_metharray_QC_meta(array.meta = "controlsEPIC")
# Merge red and green channels intensities with QC metadata
dt.mrg <- mergeQC_intensities_and_meta(
    RnBSet = rnb.set, DT.QC.meta = dt.meta)
#Draw probe specific QC plot for QC probe "21630339"
probe.plot <- plot_array_QCprobe(
    array.type = "EPIC", probe.ID = "21630339", QC.data = dt.mrg,
    DT.QC.meta = dt.meta)
#Save plot in a PDF file
ggsave(filename = "QCprobe_21630339.pdf", plot = probe.plot, device = "pdf",
       path = "~/")
}
\author{
Yoann Pageaud.
}
