% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot_asso_annot_PC}
\alias{plot_asso_annot_PC}
\title{Draws association test results between annotations from an RnBSet and PCs
from a prcomp object.}
\usage{
plot_asso_annot_PC(
  RnBSet,
  prcomp.res,
  perm.count = 10000,
  max.PCs = 8,
  PC_type.str = NULL,
  cohort.name = "dataset",
  verbose = FALSE
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation array
data and experimental quality information (Bisulfite data
not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{prcomp.res}{A PCA result of classes \code{prcomp} or
\code{irlba_prcomp} resulting from stats::prcomp() or
irlba::prcomp_irlba().}

\item{perm.count}{An \code{integer} specifying the number of permutations to
realize on a vector, for the permutations matrix
initialization, to be used for calculating the
significance of a correlation test
(Default: perm.count = 10000).}

\item{max.PCs}{An \code{integer} specifying the maximum number of
principal components to consider for association tests
with annotations (Default: max.PCs = 8).}

\item{PC_type.str}{A \code{character} string specifying what kind of data the
prcomp object has been computed from (e.g. 'QC probes',
'methylation probes', 'CpG islands', ...) to be appended
to the plot title.}

\item{cohort.name}{A \code{character} string specifying the name you want to
give to your dataset, that will be included in the plot
title (Default: cohort.name = "dataset").}

\item{verbose}{A \code{logical} to display information about the
step-by-step processing of the data if TRUE
(Default: verbose = FALSE).}
}
\value{
A \code{gg} plot.
}
\description{
Draws association test results between annotations from an RnBSet and PCs
from a prcomp object.
}
\examples{
# Create an RnBSet for Human Methylation 450K data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiData")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiData")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
# Compute PCA on QC probes from the RnBSet     
pca_res <- comp_RnBqc2PCA(RnBSet = rnb.set)
# Plot association test between annotations and PCs from QC probes results
plot_asso_annot_PC(
    RnBSet = rnb.set, prcomp.res = pca_res$prcomp, PC_type.str = "QC probes",
    cohort.name = "minfiData")
}
\author{
Yoann Pageaud.
}
