% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot_asso_annot_QC}
\alias{plot_asso_annot_QC}
\title{Draws association test results between annotations from an RnBSet and QC
probes intensities.}
\usage{
plot_asso_annot_QC(
  RnBSet,
  perm.count = 10000,
  max.QCprobes = 50,
  cohort.name = "dataset",
  verbose = FALSE,
  ncores = 1
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation
array data and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation
 array data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{perm.count}{An \code{integer} specifying the number of permutations
to realize on a vector, for the permutations matrix
initialization, to be used for calculating the
significance of a correlation test
(Default: perm.count = 10000).}

\item{max.QCprobes}{An \code{integer} specifying how many QC probes should be
kept as the top QC probes associated with RnBSet
annotations (Default: max.QCprobes = 50).}

\item{cohort.name}{A \code{character} string specifying the name you want to
give to your dataset, that will be included in the plot
title (Default: cohort.name = "dataset").}

\item{verbose}{A \code{logical} to display information about the
step-by-step processing of the data if TRUE
(Default: verbose = FALSE).}

\item{ncores}{An \code{integer} to specify the number of cores/threads
to be used to parallel-compute association tests between
annotations and QC probes intensities.}
}
\value{
A \code{gg} plot.
}
\description{
Draws association test results between annotations from an RnBSet and QC
probes intensities.
}
\examples{
# Create an RnBSet for Human Methylation 450K data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiData")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiData")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
# Plot association test between annotations and QC probes intensities results
plot_asso_annot_QC(RnBSet = rnb.set, cohort.name = "minfiData")
}
\author{
Yoann Pageaud.
}
