% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plot_negative_FFPE}
\alias{plot_negative_FFPE}
\title{Plots FFPE negative control probe fluorescence intensities barplots (WARNING: Experimental).}
\usage{
plot_negative_FFPE(RnBSet, cohort = "RnBSet")
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation array data
and experimental quality information (HM450K and Bisulfite data
not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array data
 in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{cohort}{A \code{character} string to specify the name of the cohort to
be displayed as part of the plot title
(Default: cohort = "RnBSet").}
}
\value{
A \code{gtable} barplot of the QC probe fluorescence intensities.
}
\description{
Plots FFPE negative control probe fluorescence intensities barplots (WARNING: Experimental).
}
\details{
In MethylationEPIC array we identified the Negative quality control probe
"Negative 1720" as sensitive to Formaldehyd Fixed Paraffin Embed (FFPE)
sample treatment.\cr This probe (Negative 1720; ID = 36729435) displays
unusual high fluorescence intensities in FFPE samples in both, green and red
channels, more pronounced in the green channel.\cr We suspect that a similar
Negative quality control probe also exists in HM450K array, but have not yet
identified one.
}
\examples{
# Create an RnBSet for MethylationEPIC data (May take up to 1 minute)
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
RnBeads::rnb.options(identifiers.column = "barcode")
# Plot FFPE negative control probe
neg.FFPE <- methview.qc:::plot_negative_FFPE(
    RnBSet = rnb.set, cohort = "minfiDataEPIC")
#Save plot
ggsave(filename = "FFPE_negative_control_probe_MethylationEPIC.pdf",
       plot = neg.FFPE, device = "pdf", width = 11, height = 7.3,
       path = "~/")
}
\author{
Yoann Pageaud.
}
\keyword{internal}
