% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{rnb_add_runinfo}
\alias{rnb_add_runinfo}
\title{Adds IDATs runinfo to an RnBSet pheno table.}
\usage{
rnb_add_runinfo(IDATs_dir, RnBSet)
}
\arguments{
\item{IDATs_dir}{A \code{character} specifying the directory where IDAT
files are stored that will be search recursively for
the files matching the sentrix barcode.}

\item{RnBSet}{An \code{RnBSet} basic object for storing HM450K DNA
methylation and experimental quality information (Bisulfite
data are not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}
}
\value{
An \code{RnBSet} with an updated annotation table containing runinfo
        data.
}
\description{
Adds IDATs runinfo to an RnBSet pheno table.
}
\examples{
# Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
# Specify the column containing samples identifiers (sentrix barcodes).
rnb.options(identifiers.column = 'barcode')
# Print RnBSet annotation table
pheno(rnb.set)
# Add runinfo data from the IDAT files to the RnBSet annotation table
rnb.set <- rnb_add_runinfo(IDATs_dir = idat.dir, RnBSet = rnb.set)
# Print RnBSet updated annotation table with runinfo data
pheno(rnb.set)
}
\references{
ML Smith, KA Baggerly, H Bengtsson, ME Ritchie & KD Hansen.
            illuminaio: An open source IDAT parsing tool for Illumina
            microarrays, F1000Research, (2013) 2:264.
}
\author{
Yoann Pageaud.
}
