% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{rnb_test_asso_all_annot}
\alias{rnb_test_asso_all_annot}
\title{Tests associations between all annotations in a RnBSet.}
\usage{
rnb_test_asso_all_annot(RnBSet, perm.count = 10000, verbose = FALSE)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation
array data and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation
 array data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{perm.count}{An \code{integer} specifying the number of permutations
to realize on a vector, for the permutations matrix
initialization, to be used for calculating the
significance of a correlation test
(Default: perm.count = 10000).}

\item{verbose}{A \code{logical} to display information about the
step-by-step processing of the data if TRUE
(Default: verbose = FALSE).}
}
\value{
A \code{data.table} containing all association test results.
}
\description{
Tests associations between all annotations in a RnBSet.
}
\author{
Yoann Pageaud.
}
