% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{sampleQC_crossbi}
\alias{sampleQC_crossbi}
\title{Draws a customizable PCA cross biplot on samples methylation array QC data}
\usage{
sampleQC_crossbi(
  RnBSet,
  PCs = c(1:5),
  loadings = TRUE,
  loadings.col = "blue",
  point.size = 2.5,
  top.load.by.quad = NULL,
  color.data = "ID",
  shape.data = NULL
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation
array data and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation
 array data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{PCs}{An \code{integer} vector matching principal
components to be used to generate the cross-biplot.}

\item{loadings}{A \code{logical} specifying whether the loadings
should be displayed (TRUE) or not (FALSE).}

\item{loadings.col}{A \code{character} specifying a color to be used for
loadings.}

\item{point.size}{A \code{double} specifying the size of points.}

\item{top.load.by.quad}{An \code{integer} specifying the top n most important
loadings to be displayed in the four quadrants of the
biplot graph (by quadrants). This parameters allows
to display only the most important loadings, and to
hide the less important ones, to improve visibility
when there is too many of them
(Default: top.load.by.quad = 5).}

\item{color.data}{A \code{character} specifying the column name in
'data' to be used to map colors to points. You can
specify your own custom palette of colors using the
'scale_color_manual()' function. For more information
about how to use it see
\link[ggplot2]{scale_color_manual}.}

\item{shape.data}{A \code{character} specifying the column name in
'data' to be used to map shapes to points. You can
specify your own custom set of point shapes using the
'scale_shape_manual()' function. For more information
about how to use it see
\link[ggplot2]{scale_shape_manual}.}
}
\value{
A customizable \code{gg} object of samples PCA cross biplot.
}
\description{
Draws a customizable PCA cross biplot on samples methylation array QC data
}
\examples{
# Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
RnBeads::rnb.options(identifiers.column = "barcode")
#Draw samples biplot on quality control data
sampleQC_crossbi(
    RnBSet = rnb.set, color.data = "Sample_Name", top.load.by.quad = 1)
}
\references{
Pageaud Y. et al., BiocompR - Advanced visualizations for data
            comparison.
}
\author{
Yoann Pageaud.
}
