% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_format.R
\name{smart.tolower}
\alias{smart.tolower}
\title{Convert strings to lowercase in a smart way.}
\usage{
smart.tolower(x, sep = c(" ", "_"))
}
\arguments{
\item{x}{A \code{character} string.}

\item{sep}{A \code{character} vector to specify multiple separator to remove
during lowercase conversion (Default: sep = c(" ", "_")).}
}
\value{
A \code{character} string converted in lowercase.
}
\description{
Convert strings to lowercase in a smart way.
}
\examples{
smart.tolower(x = "A_QUIET_PLACE_PART_II")
}
\author{
Yoann Pageaud.
}
\keyword{internal}
