% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{snp.heatmap}
\alias{snp.heatmap}
\title{Draws a heatmap from methylation array genotyping probes.}
\usage{
snp.heatmap(
  RnBSet,
  dist.method = "manhattan",
  annot.grps,
  annot.pal,
  heatmap.pal = c("#2166AC", "#4393C3", "#92C5DE", "#D1E5F0", "#FDDBC7", "#F4A582",
    "#D6604D", "#B2182B"),
  x.lab = "Samples",
  plot.title = NULL,
  axis.text.x = element_text(size = 10, angle = -45, hjust = 0, vjust = 0.5, color =
    "black"),
  axis.text.y.right = element_text(size = 7, color = "black"),
  axis.title.y.right = element_text(size = 11),
  lgd.text = element_text(size = 10),
  lgd.space.width = 1,
  annot.size = 1,
  dend.size = 1
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing
methylation array data and experimental quality
information (Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object
 read \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import
 methylation array data in the RnBSet see options
 available in \link[RnBeads]{rnb.options}.}
}}

\item{dist.method}{A \code{character} vector to specify the distance
methods to be used on the matrix rows and columns.
\itemize{
 \item{If the vector is of length 1: the given
       method will be applied on rows and columns
       of the matrix.}
 \item{If the vector is of length 2: the first
       method will be applied on matrix rows, and
       the second method will be applied on matrix
       columns.}
}
(Default: dist.method = 'manhattan';
Supported: dist.method = c('euclidean', 'maximum',
'manhattan', 'canberra', 'binary', 'minkowski',
'none')).}

\item{annot.grps}{A \code{list} of vectors of groups to which
variables belongs for the annotation sidebars.
Vectors' lengths have to match the number of
variables.}

\item{annot.pal}{A \code{vector} or a list of vectors containing
colors as characters for the annotation sidebars.
The length of vectors has to match the number of
levels of vectors listed in 'annot.grps'.
\itemize{
 \item{If annot.pal is a list: its length must
       match the length of the list provided to
       'annot.grps'.}
 \item{If annot.pal is a vector: make sure that the
       levels content of annotations listed in
       'annot.grps' is the same, and that no
       annotation contains less or more levels than
       another one in 'annot.grps'.}
}}

\item{heatmap.pal}{A \code{character} vector to be used as a color
palette for the heatmap.}

\item{x.lab}{A \code{character} to specify X-axis title
(Default: x.lab = 'Samples').}

\item{plot.title}{A \code{character} to be used as title for the
plot (Default: plot.title = NULL ; ).}

\item{axis.text.x}{An \code{element_text} object to setup X axis text
(Default: axis.text.x = element_text(size = 10,
angle = -45, hjust = 0, vjust = 0.5,
color = "black")).}

\item{axis.text.y.right}{An \code{element_text} object to setup right Y axis
text (Default: axis.text.y.right =
element_text(size = 7, color = "black")).}

\item{axis.title.y.right}{An \code{element_text} object to setup right Y axis
title
(Default: axis.title.y.right =
element_text(size = 11)).}

\item{lgd.text}{An \code{element_text} object to setup legend
labels (Default: lgd.text =
element_text(size = 10)).}

\item{lgd.space.width}{A \code{numeric} specifying the width of the legend
space (Default: lgd.space.width = 1).}

\item{annot.size}{A \code{numeric} defining the width of the
annotation bars (Default: annot.size = 1).}

\item{dend.size}{A \code{numeric} defining the height of the
dendrogram made on columns
(Default: dend.col.size = 1).}
}
\value{
A \code{grob} of the heatmap created on methylation array
        genotyping probes.
}
\description{
Draws a heatmap from methylation array genotyping probes.
}
\examples{
#Create an RnBSet for MethylationEPIC data.
library(RnBeads)
idat.dir <- "~/data/MethylationEPIC/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(data.source = data.source, data.type = "idat.dir")
rnb.options(identifiers.column = "barcode")
#Load scientific palette package ggsci.
library(ggsci)
#Plot heatmap of MethylationEPIC genotyping probes.
snp.htmp <- snp.heatmap(
  RnBSet = rnb.set,
  annot.grps = list("Donors" = rnb.set@pheno$ID),
  annot.pal = pal_npg("nrc", alpha = 1)(10))
#Save heatmap in a PDF file.
ggsave(
  filename = "heatmap.pdf", plot = snp.htmp$result.grob, device = "pdf",
  path = "~/")
}
\references{
Pageaud Y. et al., BiocompR - Advanced visualizations for data
            comparison.

Assenov Y. et al., Comprehensive analysis of DNA methylation data
            with RnBeads.
}
\author{
Yoann Pageaud.
}
