% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{target.biplot}
\alias{target.biplot}
\title{Draws a PCA biplots on methylation array quality control targets}
\usage{
target.biplot(
  RnBSet,
  PCx = 1,
  PCy = 2,
  point.size = 3,
  loadings = TRUE,
  loadings.col = "blue",
  top.load.by.quad = NULL
)
}
\arguments{
\item{RnBSet}{A \code{RnBSet} basic object for storing methylation
array data and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation
 array data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{PCx}{An \code{integer} matching the principal component
values to display on X-axis.}

\item{PCy}{An \code{integer} matching the principal component
values to display on Y-axis.}

\item{point.size}{A \code{double} specifying the size of points.}

\item{loadings}{A \code{logical} specifying whether the loadings
should be displayed (TRUE) or not (FALSE).}

\item{loadings.col}{A \code{character} specifying a color to be used for
loadings.}

\item{top.load.by.quad}{An \code{integer} specifying the top n most important
loadings to be displayed in the four quadrants of the
biplot graph (by quadrants). This parameters allows
to display only the most important loadings, and to
hide the less important ones, to improve visibility
when there is too many of them.}
}
\value{
A \code{gg} object of targets PCA biplot.
}
\description{
Draws a PCA biplots on methylation array quality control targets
}
\examples{
# Create an RnBSet for MethylationEPIC data (May take up to 1 minute)
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
RnBeads::rnb.options(identifiers.column = "barcode")
#Draw samples biplot on quality control data
target.biplot(RnBSet = rnb.set)
}
\references{
Pageaud Y. et al., BiocompR - Advanced visualizations for data
            comparison.
}
\author{
Yoann Pageaud.
}
