% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{update.target.meta}
\alias{update.target.meta}
\title{Updates a target intensities metadata.}
\usage{
update.target.meta(QC.data, DT.QC.meta, target, ncores = 1)
}
\arguments{
\item{QC.data}{A \code{data.table} list matching QC metadata with green
channel and red channel intensities, obtained with the
function \link{merge.QC.intensities.and.meta}.}

\item{DT.QC.meta}{A \code{data.table} with methylation array quality control
metadata obtained with the function
\link{load.metharray.QC.meta}.}

\item{target}{A \code{character} specifying the name of the target step
for which the metadata should be updated.}

\item{ncores}{An \code{integer} specifying the number of cores or threads
to be used for parallel processing.}
}
\value{
A \code{data.table} containing the updated metadata for a given
        target step of a given quality control dataset.
}
\description{
Updates a target intensities metadata.
}
\examples{
update.target.meta(QC.data = QC.data, target = "Hybridization")
}
\author{
Yoann Pageaud.
}
\keyword{internal}
