% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{update_target_meta}
\alias{update_target_meta}
\title{Updates a target intensities metadata.}
\usage{
update_target_meta(QC.data, DT.QC.meta, target, ncores = 1)
}
\arguments{
\item{QC.data}{A \code{data.table} list matching QC metadata with green
channel and red channel intensities, obtained with the
function \link{mergeQC_intensities_and_meta}.}

\item{DT.QC.meta}{A \code{data.table} with methylation array quality control
metadata obtained with the function
\link{load_metharray_QC_meta}.}

\item{target}{A \code{character} specifying the name of the target step
for which the metadata should be updated.}

\item{ncores}{An \code{integer} specifying the number of cores or threads
to be used for parallel processing.}
}
\value{
A \code{data.table} containing the updated metadata for a given
        target step of a given quality control dataset.
}
\description{
Updates a target intensities metadata.
}
\examples{
#Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
# Specify the column containing samples identifiers (sentrix barcodes).
RnBeads::rnb.options(identifiers.column = "barcode")
#Create the data.table with quality control metadata
dt.meta <- load_metharray_QC_meta(array.meta = "controlsEPIC")
# Merge red and green channels intensities with QC metadata
dt.mrg <- mergeQC_intensities_and_meta(
   RnBSet = rnb.set, DT.QC.meta = dt.meta)
update_target_meta(QC.data = QC.data, target = "Hybridization")
}
\author{
Yoann Pageaud.
}
\keyword{internal}
