% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{cohort.gp.density}
\alias{cohort.gp.density}
\title{Displays the distribution of genotyping probes values in a cohort.}
\usage{
cohort.gp.density(RnB.set)
}
\arguments{
\item{RnB.set}{A \code{RnBSet} basic object for storing methylation array
data and experimental quality information (HM450K and
Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}
}
\value{
A \code{gg} plot of the results.
}
\description{
Displays the distribution of genotyping probes values in a cohort.
}
\examples{
#Create an RnBSet for MethylationEPIC data
library(RnBeads)
idat.dir <- "~/data/MethylationEPIC/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(
  data.source = data.source, data.type = "idat.dir")
rnb.options(identifiers.column = "barcode")
# Draw the genotyping probes values offset plot
methview.qc::cohort.gp.density(RnB.set = rnb.set)
}
\author{
Yoann Pageaud.
}
