% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{devscore.fluo}
\alias{devscore.fluo}
\title{Computes a deviation score between samples fluorescence and an internal
HM450K reference.}
\usage{
devscore.fluo(RnBSet, samples, target, ncores = 1)
}
\arguments{
\item{RnBSet}{An \code{RnBSet} basic object for storing HM450K DNA
methylation and experimental quality information (Bisulfite
data  and MethylationEPIC data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{samples}{A \code{character} vector specifying the samples to include
for the deviation score calculation. You can catch the sample
IDs you wish to evaluate running \code{RnBSet@pheno[,1]}}

\item{target}{A \code{character} specifying the name of the target step to
evaluate fluorescence from.}

\item{ncores}{An \code{integer} specifying the number of cores or threads to
be used for parallel processing.}
}
\value{
A \code{data.table} with probes data and samples computed deviation
        scores.
}
\description{
Computes a deviation score between samples fluorescence and an internal
HM450K reference.
}
\details{
The deviation score is calculated as following:
         \enumerate{
          \item{the square root of samples intensities fluorescence is
          calculated for both red & green channels.}
          \item{the square root of the reference intensities fluorescence is
          calculated the same way.}
          \item{Then the ratio of the samples values out of the reference
          values is calculated.}
          \item{The previous result is multiplied by 100 to get a percentage.
          }
          \item{Then 100 is substracted to the result in order to obtain the
          delta between this percentage and the reference (which counts as
          100\% signal intensity). The resulting values are deviation scores
          of fluorescence computed for each probes and each samples. It can
          be extracted from the resulting data.table in the column
          'percent.diff.sqrt'.}
         }
}
\examples{
#Create an RnBSet for HM450K data
library(RnBeads)
idat.dir <- "~/data/my_idat_dir/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(
  data.source = data.source, data.type = "idat.dir")
#Compute deviation score
devscore.fluo(RnBSet = rnb.set, samples = c("13169","13947","14312","14359"),
              target = "Hybridization")
}
\author{
Yoann Pageaud.
}
