% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{merge.QC.intensities.and.meta}
\alias{merge.QC.intensities.and.meta}
\title{Merges red and green channels intensities with QC probes metadata.}
\usage{
\method{merge}{QC.intensities.and.meta}(RnBSet, DT.QC.meta)
}
\arguments{
\item{RnBSet}{An \code{RnBSet} basic object for storing methylation
array DNA methylation and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{DT.QC.meta}{A \code{data.table} with methylation array quality control
metadata obtained using \link{load.metharray.QC.meta}.}
}
\value{
A \code{data.table} list matching QC metadata with green channel and
        red channel intensities.
}
\description{
Merges red and green channels intensities with QC probes metadata.
}
\examples{
#Create an RnBSet for MethylationEPIC data
library(RnBeads)
idat.dir <- "~/data/MethylationEPIC/"
sample.annotation <- "~/data/Annotations/sample_sheet.csv"
data.source <- c(idat.dir, sample.annotation)
rnb.set <- rnb.execute.import(data.source = data.source, data.type = "idat.dir")
rnb.options(identifiers.column = "barcode")
#Create the data.table with quality control metadata
dt.meta <- load.metharray.QC.meta(array.meta = "controlsEPIC")
# Merge red and green channels intensities with QC metadata
dt.mrg <- merge.QC.intensities.and.meta(RnBSet = rnb.set, DT.QC.meta = dt.meta)
}
\author{
Yoann Pageaud.
}
