% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fun.R
\name{RnB2PCA}
\alias{RnB2PCA}
\title{Computes a PCA from an RnBSet on a subset of selected probes.}
\usage{
RnB2PCA(RnBSet, probe.type = "cg", nPCs = NULL, scaling = FALSE)
}
\arguments{
\item{RnBSet}{An \code{RnBSet} basic object for storing methylation
array DNA methylation and experimental quality information
(Bisulfite data not supported).
\itemize{
 \item{For more information about RnBSet object read
 \link[RnBeads]{RnBSet-class}.}
 \item{To create an RnBSet object run
 \link[RnBeads]{rnb.execute.import}.}
 \item{For additionnal options to import methylation array
 data in the RnBSet see options available in
 \link[RnBeads]{rnb.options}.}
}}

\item{probe.type}{A \code{character} to specify the type of probes on which
the principal component should be computed
(Default: probe.type = 'cg';
Supported: probe.type = c('cg', 'ch', 'rs', 'qc')).
\itemize{
 \item{cg - CpG methylation probes beta values.}
 \item{ch - CpH (Cytosine di-/tri-nucleotide) methylation
       probes beta values.}
 \item{rs - SNPs genotyping probes allelic version.}
 \item{qc - Quality control probes fluorescence
       intensities.}
}}

\item{nPCs}{An \code{integer} specifying the number of principal
components you wish to compute on your RnBSet
(Default: nPCs = NULL will compute all principal
components).}
}
\value{
A \code{list} containing a prcomp object with all results from the
        PCA, and a data.table with all the RnBSet data.
}
\description{
Computes a PCA from an RnBSet on a subset of selected probes.
}
\examples{
# Create an RnBSet for MethylationEPIC data
require(Biobase)
idat.dir <- system.file("extdata", package = "minfiDataEPIC")
sample.annotation <- system.file(
    "extdata", "SampleSheet.csv", package = "minfiDataEPIC")
data.source <- c(idat.dir, sample.annotation)
rnb.set <- RnBeads::rnb.execute.import(
    data.source = data.source, data.type = "idat.dir")
# Compute PCA on CG methylation probes intensities from the RnBSet
RnB2PCA(RnBSet = rnb.set) # By default computes only top 10 PCs  
}
\author{
Yoann Pageaud.
}
